implement customerController
